+++ src/Admin/ProcedureAdmin.php
@@ -13,6 +13,7 @@ use Sulu\Bundle\AdminBundle\Admin\View\TogglerToolbarAction;
 use Sulu\Bundle\AdminBundle\Admin\View\ToolbarAction;
 use Sulu\Bundle\AdminBundle\Admin\View\ViewBuilderFactoryInterface;
 use Sulu\Bundle\AdminBundle\Admin\View\ViewCollection;
+use Sulu\Bundle\ReferenceBundle\Infrastructure\Sulu\Admin\View\ReferenceViewBuilderFactoryInterface;
 use Sulu\Component\Security\Authorization\PermissionTypes;
 use Sulu\Component\Security\Authorization\SecurityCheckerInterface;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
@@ -39,16 +40,20 @@ class ProcedureAdmin extends Admin
 
     private ActivityViewBuilderFactoryInterface $activityViewBuilderFactory;
 
+    private ReferenceViewBuilderFactoryInterface $referenceViewBuilderFactory;
+
     public function __construct(
         ViewBuilderFactoryInterface $viewBuilderFactory,
         SecurityCheckerInterface $securityChecker,
         WebspaceManagerInterface $webspaceManager,
-        ActivityViewBuilderFactoryInterface $activityViewBuilderFactory
+        ActivityViewBuilderFactoryInterface $activityViewBuilderFactory,
+        ReferenceViewBuilderFactoryInterface $referenceViewBuilderFactory
     ) {
         $this->viewBuilderFactory = $viewBuilderFactory;
         $this->securityChecker = $securityChecker;
         $this->webspaceManager = $webspaceManager;
         $this->activityViewBuilderFactory = $activityViewBuilderFactory;
+        $this->referenceViewBuilderFactory = $referenceViewBuilderFactory;
     }
 
     public function configureNavigationItems(NavigationItemCollection $navigationItemCollection): void
@@ -151,6 +156,13 @@ class ProcedureAdmin extends Admin
             if ($this->activityViewBuilderFactory->hasActivityListPermission()) {
                 $viewCollection->add(
                     $this->activityViewBuilderFactory->createActivityListViewBuilder(static::EDIT_FORM_VIEW . ".activity", "/activity", Procedure::RESOURCE_KEY)
+                        ->setParent(static::EDIT_FORM_VIEW)
+                );
+            }
+
+            if ($this->referenceViewBuilderFactory->hasReferenceListPermission()) {
+                $viewCollection->add(
+                    $this->referenceViewBuilderFactory->createReferenceListViewBuilder(static::EDIT_FORM_VIEW . ".insights.reference", "/references", Procedure::RESOURCE_KEY)
                         ->setParent(static::EDIT_FORM_VIEW)
                 );
             }