+++ src/Content/Type/SinglePublicMarketSelection.php
@@ -6,10 +6,12 @@ namespace Pixel\TownHallBundle\Content\Type;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Pixel\TownHallBundle\Entity\PublicMarket;
+use Sulu\Bundle\ReferenceBundle\Application\Collector\ReferenceCollectorInterface;
+use Sulu\Bundle\ReferenceBundle\Infrastructure\Sulu\ContentType\ReferenceContentTypeInterface;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\SimpleContentType;
 
-class SinglePublicMarketSelection extends SimpleContentType
+class SinglePublicMarketSelection extends SimpleContentType implements ReferenceContentTypeInterface
 {
     protected EntityManagerInterface $entityManager;
 
@@ -35,5 +37,19 @@ class SinglePublicMarketSelection extends SimpleContentType
         return [
             'ids' => $property->getValue(),
         ];
+    }
+
+    public function getReferences(PropertyInterface $property, ReferenceCollectorInterface $referenceCollector, string $propertyPrefix = ''): void
+    {
+        $data = $property->getValue();
+        if (! isset($data) || ! is_int($data)) {
+            return;
+        }
+
+        $referenceCollector->addReference(
+            PublicMarket::RESOURCE_KEY,
+            (string) $data,
+            $propertyPrefix . $property->getName()
+        );
     }
 }